backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
distlib==0.3.6
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.1.0
filelock==3.8.0
frozenlist==1.3.3
h11==0.12.0
httpcore==0.15.0
httpx==0.24.1
idna==3.6
iniconfig==1.1.1
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.5
multidict==6.0.4
murmurhash==1.0.9
numpy==1.26.4
openai==0.27.2
packaging==22.0
pandas==2.2.0
pathy==0.10.1
patsy==0.5.3
platformdirs==2.5.2
pluggy==1.0.0
preshed==3.0.8
pydantic==1.10.7
PyJWT==2.6.0
pysonic_channel==0.1.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.4
pytz==2024.1
pywin32==304
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
statsmodels==0.13.5
thinc==8.1.9
tomli==2.0.1
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.9.0
tzdata==2022.7
urllib3==2.1.0
virtualenv==20.16.5
wasabi==1.1.1
yarl==1.8.2
